requirements.txt
# Core Flask dependencies
Flask>=2.3.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# AI/ML
openai>=1.0.0
tiktoken>=0.5.0

# Document processing
pandas>=2.0.0
python-docx>=0.8.0
PyPDF2>=3.0.0
openpyxl>=3.1.0

# Development dependencies (install with: pip install -r requirements.txt -r requirements-dev.txt)
# Or uncomment for development:
# pytest>=7.0.0
# pytest-flask>=1.2.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0